backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
mysqlclient>=2.2.7
python-dotenv>=1.0.0
djangorestframework-simplejwt==5.3.1
requests>=2.31.0
channels>=4.0.0
channels-redis>=4.1.0
